# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HalkYemek API
  version: dev
  description: Beylikdüzü odaklı QR + Wallet yemek sistemi
paths:
...
  /api/v1/businesses/{business_id}/checkout-sessions/{token}/consume/:
    post:
      operationId: checkout_session_consume
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - checkout
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutConsumeResponse'
              examples:
                StandardErrorEnvelope:
                  value:
                    ok: false
                    error:
                      code: ValidationError
                      message:
                        amount:
                        - Ensure this value is greater than or equal to 1.
                      request_id: 9b6b2d4cc92f44a28e3a5d2c7d92b001
                  summary: Standard error envelope
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '410':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
  /api/v1/businesses/{business_id}/checkout-sessions/{token}/preview/:
    get:
      operationId: checkout_session_consume_preview
      parameters:
      - in: path
        name: business_id
        schema:
          type: integer
        required: true
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - checkout
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionPreviewResponse'
              examples:
                StandardErrorEnvelope:
                  value:
                    ok: false
                    error:
                      code: ValidationError
                      message:
                        amount:
                        - Ensure this value is greater than or equal to 1.
                      request_id: 9b6b2d4cc92f44a28e3a5d2c7d92b001
                  summary: Standard error envelope
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '410':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
...
  /api/v1/checkout-sessions/:
    post:
      operationId: checkout_session_create
      tags:
      - checkout
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionDetail'
              examples:
                CheckoutSessionCreated:
                  value:
                    id: 321
                    token: chk_live_token
                    status: PENDING
                    amount: 215
                    subtotal_amount: 200
                    customer_fee_amount: 15
                    business_fee_amount: 10
                    business_net_amount: 190
                    platform_total_fee_amount: 25
                    item_count: 2
                    currency: TRY
                    expires_at: '2026-04-02T13:00:00+03:00'
                    business:
                      id: 8
                      name: Örnek İşletme
                    cart:
                      id: 55
                    pricing:
                      fee_model: customer_fee
                    items:
                    - menu_item_id: 34
                      menu_item_name: Mercimek Çorbası
                      quantity: 2
                      unit_price_amount: 100
                      line_total_amount: 200
                      sort_order: 1
                  summary: Checkout session created
                StandardErrorEnvelope:
                  value:
                    ok: false
                    error:
                      code: ValidationError
                      message:
                        amount:
                        - Ensure this value is greater than or equal to 1.
                      request_id: 9b6b2d4cc92f44a28e3a5d2c7d92b001
                  summary: Standard error envelope
          description: ''
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
  /api/v1/checkout-sessions/{token}/:
    get:
      operationId: checkout_session_detail
      parameters:
      - in: path
        name: token
        schema:
          type: string
        required: true
      tags:
      - checkout
      security:
      - jwtAuth: []
      - bearerAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CheckoutSessionDetail'
              examples:
                StandardErrorEnvelope:
                  value:
                    ok: false
                    error:
                      code: ValidationError
                      message:
                        amount:
                        - Ensure this value is greater than or equal to 1.
                      request_id: 9b6b2d4cc92f44a28e3a5d2c7d92b001
                  summary: Standard error envelope
          description: ''
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
        '410':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorEnvelope'
          description: ''
...
components:
  schemas:
...
    ApiErrorBody:
      type: object
      properties:
        code:
          type: string
        message: {}
        request_id:
          type: string
        details: {}
        reason:
          type: string
      required:
      - code
      - message
      - request_id
    ApiErrorEnvelope:
      type: object
      properties:
        ok:
          type: boolean
          default: false
        error:
          $ref: '#/components/schemas/ApiErrorBody'
      required:
      - error
...
    CheckoutConsumeResponse:
      type: object
      properties:
        status:
          type: string
        order_id:
          type: integer
        amount:
          type: integer
          description: Deprecated mirror of total_charged_amount.
        total_charged_amount:
          type: integer
        checkout_session_id:
          type: integer
      required:
      - amount
      - checkout_session_id
      - order_id
      - status
      - total_charged_amount
    CheckoutSessionBusiness:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
      required:
      - id
      - name
    CheckoutSessionDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        token:
          type: string
          maxLength: 64
        status:
          $ref: '#/components/schemas/CheckoutSessionDetailStatusEnum'
        amount:
          type: integer
          readOnly: true
          description: Deprecated mirror of total_payable_amount. Frontend must use
            total_payable_amount.
        total_payable_amount:
          type: integer
          readOnly: true
        subtotal_amount:
          type: integer
        customer_fee_amount:
          type: integer
        business_fee_amount:
          type: integer
        business_net_amount:
          type: integer
        platform_total_fee_amount:
          type: integer
        item_count:
          type: integer
        currency:
          type: string
          maxLength: 8
        expires_at:
          type: string
          format: date-time
        business:
          type: object
          additionalProperties: {}
          readOnly: true
        cart:
          type: object
          additionalProperties: {}
          readOnly: true
        pricing:
          type: object
          additionalProperties: {}
          readOnly: true
        items:
          type: array
          items:
            type: object
            additionalProperties: {}
          readOnly: true
      required:
      - amount
      - business
      - cart
      - expires_at
      - id
      - items
      - pricing
      - token
      - total_payable_amount
    CheckoutSessionDetailStatusEnum:
      enum:
      - PENDING
      - CONFIRMED
      - CONSUMED
      - EXPIRED
      - CANCELLED
      type: string
      description: |-
        * `PENDING` - Pending
        * `CONFIRMED` - Confirmed
        * `CONSUMED` - Consumed
        * `EXPIRED` - Expired
        * `CANCELLED` - Cancelled
    CheckoutSessionPreviewResponse:
      type: object
      properties:
        checkout_session_id:
          type: integer
        token:
          type: string
        status:
          type: string
        expires_at:
          type: string
          format: date-time
          nullable: true
        amount:
          type: integer
          description: Deprecated mirror of total_payable_amount.
        total_payable_amount:
          type: integer
        item_count:
          type: integer
        business:
          $ref: '#/components/schemas/CheckoutSessionBusiness'
        can_consume:
          type: boolean
        failure_reason:
          type: string
        existing_order_id:
          type: integer
          nullable: true
      required:
      - amount
      - business
      - can_consume
      - checkout_session_id
      - existing_order_id
      - expires_at
      - failure_reason
      - item_count
      - status
      - token
      - total_payable_amount
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...